Break apart and reorganize ActionController instrumentation
