+++ lib/traceview/frameworks/rails/inst/action_controller.rb
@@ -69,145 +69,12 @@ module TraceView
         TraceView::API.log_exit('actionview')
       end
     end
-
-    #
-    # ActionController3
-    #
-    # This modules contains the instrumentation code specific
-    # to Rails v3
-    #
-    module ActionController3
-      include ::TraceView::Inst::RailsBase
-
-      def self.included(base)
-        base.class_eval do
-          alias_method_chain :process, :traceview
-          alias_method_chain :process_action, :traceview
-          alias_method_chain :render, :traceview
-        end
-      end
-
-      def process_with_traceview(*args)
-        TraceView::API.log_entry('rails')
-        process_without_traceview(*args)
-
-      rescue Exception => e
-        TraceView::API.log_exception(nil, e) if log_rails_error?(e)
-        raise
-      ensure
-        TraceView::API.log_exit('rails')
-      end
-
-      def process_action_with_traceview(*args)
-        report_kvs = {
-          :Controller   => self.class.name,
-          :Action       => action_name,
-        }
-        TraceView::API.log(nil, 'info', report_kvs)
-
-        process_action_without_traceview(*args)
-      rescue Exception
-        report_kvs[:Status] = 500
-        TraceView::API.log(nil, 'info', report_kvs)
-        raise
-      end
-    end
-
-    #
-    # ActionController4
-    #
-    # This modules contains the instrumentation code specific
-    # to Rails v4
-    #
-    module ActionController4
-      include ::TraceView::Inst::RailsBase
-
-      def self.included(base)
-        base.class_eval do
-          alias_method_chain :process_action, :traceview
-          alias_method_chain :render, :traceview
-        end
-      end
-
-      def process_action_with_traceview(method_name, *args)
-        report_kvs = {
-          :Controller   => self.class.name,
-          :Action       => self.action_name,
-        }
-
-        TraceView::API.log_entry('rails', report_kvs)
-        process_action_without_traceview(method_name, *args)
-
-      rescue Exception => e
-        TraceView::API.log_exception(nil, e) if log_rails_error?(e)
-        raise
-      ensure
-        TraceView::API.log_exit('rails')
-      end
-    end
   end
 end
 
 if defined?(ActionController::Base) && TraceView::Config[:action_controller][:enabled]
-  if ::Rails::VERSION::MAJOR == 4
-
-    class ActionController::Base
-      include TraceView::Inst::ActionController4
-    end
-
-  elsif ::Rails::VERSION::MAJOR == 3
-
-    class ActionController::Base
-      include TraceView::Inst::ActionController3
-    end
-
-  elsif ::Rails::VERSION::MAJOR == 2
-
-    ActionController::Base.class_eval do
-      include ::TraceView::Inst::RailsBase
-
-      alias :perform_action_without_traceview :perform_action
-      alias :rescue_action_without_traceview :rescue_action
-      alias :process_without_traceview :process
-      alias :render_without_traceview :render
-
-      def process(*args)
-        TraceView::API.log_entry('rails')
-        process_without_traceview(*args)
-
-      rescue Exception => e
-        TraceView::API.log_exception(nil, e) if log_rails_error?(e)
-        raise
-      ensure
-        TraceView::API.log_exit('rails')
-      end
-
-      def perform_action(*arguments)
-        report_kvs = {
-          :Controller  => @_request.path_parameters['controller'],
-          :Action      => @_request.path_parameters['action']
-        }
-        TraceView::API.log(nil, 'info', report_kvs)
-        perform_action_without_traceview(*arguments)
-      end
-
-      def rescue_action(exn)
-        TraceView::API.log_exception(nil, exn) if log_rails_error?(exn)
-        rescue_action_without_traceview(exn)
-      end
-
-      def render(options = nil, extra_options = {}, &block)
-        TraceView::API.log_entry('actionview')
-        render_without_traceview(options, extra_options, &block)
-
-      rescue Exception => e
-        TraceView::API.log_exception(nil, e) if log_rails_error?(e)
-        raise
-      ensure
-        TraceView::API.log_exit('actionview')
-      end
-    end
-  end
   TraceView.logger.info '[traceview/loading] Instrumenting actioncontroler' if TraceView::Config[:verbose]
+  require "traceview/frameworks/rails/inst/action_controller#{Rails::VERSION::MAJOR}"
+  ::TraceView::Util.send_include(::ActionController::Base, TraceView::Inst::ActionController)
 end
 # vim:set expandtab:tabstop=2
